+++ lib/numbers_and_words/strategies/figures_converter/languages/nl.rb
@@ -11,6 +11,7 @@ module NumbersAndWords
 
           def capacity_iteration
             return super if FiguresArray::THOUSAND_CAPACITY != @current_capacity
+            return [] if apply_tens_of_hundreds?
             return megs if figures.number_in_capacity(@current_capacity) == 1
 
             capacity_words = words_in_capacity(@current_capacity)
@@ -23,6 +24,24 @@ module NumbersAndWords
             else
               super
             end
+          end
+
+          def number_without_capacity_to_words
+            return super unless apply_tens_of_hundreds?
+
+            base = save_parent_figures do
+              @figures = @parent_figures[2, 2].to_figures
+              simple_number_to_words
+            end
+            [([base, translate(:hundreds, 1)] + simple_number_to_words).join]
+          end
+
+          private
+
+          def apply_tens_of_hundreds?
+            @options.tens_of_hundreds.result &&
+              (1..9).cover?(figures.number_in_capacity(FiguresArray::THOUSAND_CAPACITY)) &&
+              figures.hundreds
           end
         end
       end