+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   * Drop support for Ruby 2.5. \[[#180](https://github.com/kslazarev/numbers_and_words/pull/180)\]
   * Your contribution here.
 
+### Bugs
+  * Properly handle Dutch year numbers. \[[#183](https://github.com/kslazarev/numbers_and_words/pull/183)\] \([@mvz](https://github.com/mvz)\)
+
 ## 0.11.11 (July 5, 2021)
 
 ### Bugs